requirements.txt
google-cloud-speech
google-cloud-storage
wheel
flask
gunicorn
